+++ src/car.h
@@ -28,10 +28,10 @@ class car
     void  right();
     void  stop();
     int   getDistance();
-    void  forwardT(int speed, int time);
-    void  backwardT(int speed, int time);
-    void  leftT(int speed, int time);
-    void  rightT(int speed, int time);
+    void  forwardT(int speed, float time);
+    void  backwardT(int speed, float time);
+    void  leftT(int speed, float time);
+    void  rightT(int speed, float time);
   private:
 };
